test: make cypress tests work with Archimedes
Since the code so far is salvaged from Euclid, the Cypress
tests were expecting Euclid stuff to work. This removes
those expectations, however the actual Euclid components still
need to be removed; I am leaving them in for now as I want to
use them as reference for building Archimedes.
