+++ src/client/src/screens/App/screens/Home/index.tsx
@@ -1,12 +1,10 @@
 import React from 'react';
 
 import Header from 'App/shared/components/Header';
-import ChartGroup from './components/ChartGroup';
 
 const Home = () => (
   <>
     <Header />
-    <ChartGroup />
   </>
 );
