+++ src/client/src/screens/App/shared/components/Header.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
-import ErrorBoundary from 'react-error-boundary';
 import styled from 'styled-components';
 
 import textColor from 'App/shared/styles/text-color';
-import SelectTickerSearchField from '../containers/SelectTickerSearchField';
 
 const Head = styled.section`
   text-align: center;
@@ -19,23 +17,9 @@ const Title = styled.span`
   font-size: 3rem;
 `;
 
-const SearchArea = styled.div`
-  padding-left: 30rem;
-`;
-
 const Header = () => (
   <Head>
-    <Title>Euclid</Title>
-
-    <SearchArea>
-      <ErrorBoundary>
-        {
-          // disabling because the react-redux types are broken
-          // tslint:disable-next-line no-unsafe-any
-          <SelectTickerSearchField />
-        }
-      </ErrorBoundary>
-    </SearchArea>
+    <Title>Archimedes</Title>
   </Head>
 );
